Added CLI view. Removed unused assets and png images in the root folder. Deleted old Telemetry controller and UI. Fixed issue when switching to emoji keyboard.  Fixed selection problems in config.

+++ Cleanflight Configurator/ConfigurationViewController.swift
@@ -209,7 +209,7 @@ class ConfigurationViewController: UITableViewController, FlightDataListener, UI
                                                 self.msp.sendMessage(.MSP_SET_REBOOT, data: nil, retry: 2, callback: { success in
                                                     if success {
                                                         // Wait 1500 ms
-                                                        dispatch_after(dispatch_time(DISPATCH_TIME_NOW, Int64(1500 * Double(NSEC_PER_MSEC))), dispatch_get_main_queue(), {
+                                                        dispatch_after(dispatch_time(DISPATCH_TIME_NOW, Int64(UInt64(1500) * NSEC_PER_MSEC)), dispatch_get_main_queue(), {
                                                             self.saveButton.enabled = true
                                                             appDelegate.startTimer()
                                                             // Refetch information from FC
